+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cube-rs"
+name = "cube_rs"
 version = "0.1.0"
 edition = "2021"
